Refactor 14 - interfaces package
As multiple interfaces have been created, its better for classification purposes to group them together, thus the creation of this package. Within includes :
- Collidable
- Destructible
- GameStartEnd (was Level)
- InitialiseActors (NEW)
- Movable
- Shield

classes affected by these changes have also been modified accordingly
