+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/interfaces/Collidable.java
@@ -1,4 +1,4 @@
-package dev.ccr.dmscw2024.levels;
+package dev.ccr.dmscw2024.interfaces;
 
 public interface Collidable {
     boolean checkCollision(Collidable other);
