+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/interfaces/GameStartEnd.java
@@ -1,0 +1,13 @@
+package dev.ccr.dmscw2024.interfaces;
+
+import dev.ccr.dmscw2024.levels.LevelView;
+import javafx.scene.Scene;
+
+public interface GameStartEnd {
+    Scene initializeScene();
+
+    void startGame();
+    void endGame();
+
+    LevelView instantiateLevelView();
+}
