+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/interfaces/InitialiseActors.java
@@ -1,0 +1,6 @@
+package dev.ccr.dmscw2024.interfaces;
+
+public interface InitialiseActors {
+    void initialiseFriendlyUnits();
+    void spawnEnemyUnits();
+}
