+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/interfaces/Movable.java
@@ -1,4 +1,4 @@
-package dev.ccr.dmscw2024.planes;
+package dev.ccr.dmscw2024.interfaces;
 
 public interface Movable {
     void updatePosition();
