+++ CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/interfaces/Shield.java
@@ -1,4 +1,4 @@
-package dev.ccr.dmscw2024.specials.shield;
+package dev.ccr.dmscw2024.interfaces;
 
 public interface Shield {
     void showShield();
