--- CW2024-master/CW2024-master/src/main/java/dev/ccr/dmscw2024/levels/Level.java
@@ -1,9 +1,0 @@
-package dev.ccr.dmscw2024.levels;
-
-import javafx.scene.Scene;
-
-public interface Level {
-    Scene initializeScene();
-    void startGame();
-    void endGame();
-}
