+++ CW2024-master/CW2024-master/src/main/java/module-info.java
@@ -16,4 +16,5 @@ module com.example.demo {
     opens dev.ccr.dmscw2024.projectile to javafx.fxml;
     opens dev.ccr.dmscw2024.planes to javafx.fxml;
     opens dev.ccr.dmscw2024.utility to javafx.fxml;
+    opens dev.ccr.dmscw2024.interfaces to javafx.fxml;
 }
